Extract memory CLI command
